refactor(create):  instance creation stages
- Simplified the `CreateView` component by removing unused imports and consolidating logic.
- Added new `IdentityStage` and `SetupStage` components to handle specific steps in the instance creation process.
- Introduced shared utilities for managing stages and rendering components, enhancing code organization.
- Implemented a hover prefetching mechanism for improved performance during loader selection.
- Created a view model for version management, streamlining the display of available Minecraft versions.
- Updated styling and layout for a more intuitive user experience during instance setup.
